@@ -13,8 +13,8 @@ plugins {
     id 'jacoco'
 }
 
-sourceCompatibility = 1.8
-targetCompatibility = 1.8
+sourceCompatibility = 10
+targetCompatibility = 10
 
 repositories {
     mavenCentral()
@@ -26,6 +26,7 @@ dependencies {
     compile 'org.springframework.boot:spring-boot-starter-data-jpa'
     compile 'org.springframework.boot:spring-boot-starter-security'
     compile 'org.springframework.security.oauth:spring-security-oauth2'
+    compile 'javax.annotation:javax.annotation-api:1.3.2'
     runtime 'javax.xml.bind:jaxb-api:2.3.0'
     // Vaadin
     compile 'com.vaadin:vaadin-spring-boot-starter'
