@@ -1,11 +1,10 @@
 /*
- * Copyright (C) 2016 Alice on Sunday Nights Workshop Participants. All rights reserved.
+ * Copyright (C) 2016-2018 Alice on Sunday Nights Workshop Participants. All rights reserved.
  */
 package io.github.aosn.mosaic.controller;
 
 import io.github.aosn.mosaic.ui.MainUI;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.web.authentication.logout.SecurityContextLogoutHandler;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -30,14 +29,14 @@ public class UserController {
     @GetMapping(value = {"/user", "/me"})
     public Map<String, String> user(Principal principal) {
         log.info("user(" + principal + ")");
-        Map<String, String> map = new LinkedHashMap<>();
+        var map = new LinkedHashMap<String, String>();
         map.put("name", principal.getName());
         return map;
     }
 
     @GetMapping(value = LOGOUT_PATH)
     public String logoutPage(HttpServletRequest request, HttpServletResponse response) {
-        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        var auth = SecurityContextHolder.getContext().getAuthentication();
         if (auth != null) {
             new SecurityContextLogoutHandler().logout(request, response, auth);
         }
