@@ -5,7 +5,6 @@ package io.github.aosn.mosaic.domain.repository.auth;
 
 import io.github.aosn.mosaic.domain.model.auth.GitHubOrganization;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.ResponseEntity;
 import org.springframework.security.oauth2.client.OAuth2RestTemplate;
 import org.springframework.stereotype.Repository;
 
@@ -30,11 +29,8 @@ public class GitHubOrganizationRepository {
      * @return {@link List} of {@link GitHubOrganization}s
      */
     public List<GitHubOrganization> getAll(OAuth2RestTemplate restTemplate) {
-        if (restTemplate == null) {
-            throw new NullPointerException("restTemplate is null.");
-        }
         log.info("GET " + restTemplate.getUriTemplateHandler().expand(RESOURCE_PATH));
-        ResponseEntity<GitHubOrganization[]> entity = restTemplate.getForEntity(RESOURCE_PATH, GitHubOrganization[].class);
+        var entity = restTemplate.getForEntity(RESOURCE_PATH, GitHubOrganization[].class);
         if (!entity.getStatusCode().is2xxSuccessful()) {
             log.error("GitHub error: " + entity.getStatusCodeValue());
             throw new RuntimeException("GitHub error: " + entity.getStatusCodeValue());
