@@ -1,14 +1,12 @@
 /*
- * Copyright (C) 2016 Alice on Sunday Nights Workshop Participants. All rights reserved.
+ * Copyright (C) 2016-2018 Alice on Sunday Nights Workshop Participants. All rights reserved.
  */
 package io.github.aosn.mosaic.domain.service.notification;
 
 import io.github.aosn.mosaic.domain.model.notification.SlackMessage;
-import io.github.aosn.mosaic.domain.model.poll.Group;
 import io.github.aosn.mosaic.domain.model.poll.Poll;
 import io.github.aosn.mosaic.domain.repository.notification.SlackRepository;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
 
@@ -23,17 +21,16 @@ public class NotificationService {
 
     private final SlackRepository slackRepository;
 
-    @Autowired
     public NotificationService(SlackRepository slackRepository) {
         this.slackRepository = slackRepository;
     }
 
     public void notifyBeginOfPoll(Poll poll) {
-        Group group = poll.getGroup();
+        var group = poll.getGroup();
         if (!group.isSlackEnabled()) {
             return;
         }
-        SlackMessage message = SlackMessage.builder()
+        var message = SlackMessage.builder()
                 .channel(group.getSlackChannel())
                 .username(group.getSlackUsername())
                 .text(group.getSlackBeginTemplate().replace("%s", poll.getSubject()))
@@ -42,11 +39,11 @@ public class NotificationService {
     }
 
     public void notifyEndOfPoll(Poll poll) {
-        Group group = poll.getGroup();
+        var group = poll.getGroup();
         if (!group.isSlackEnabled()) {
             return;
         }
-        SlackMessage message = SlackMessage.builder()
+        var message = SlackMessage.builder()
                 .channel(group.getSlackChannel())
                 .username(group.getSlackUsername())
                 .text(group.getSlackEndTemplate().replace("%s", poll.getSubject()))
