@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2017 Alice on Sunday Nights Workshop Participants. All rights reserved.
+ * Copyright (C) 2017-2018 Alice on Sunday Nights Workshop Participants. All rights reserved.
  */
 package io.github.aosn.mosaic.ui.view;
 
@@ -26,7 +26,6 @@ import io.github.aosn.mosaic.ui.view.layout.ViewRoot;
 import io.github.aosn.mosaic.ui.view.layout.VisibilityIndicator;
 import lombok.extern.slf4j.Slf4j;
 import org.pegdown.PegDownProcessor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.vaadin.spring.i18n.I18N;
 
 import java.text.DateFormat;
@@ -57,7 +56,6 @@ public class BookView extends CustomComponent implements View {
     private transient final PollService pollService;
     private transient final StockService stockService;
 
-    @Autowired
     public BookView(I18N i18n, UserService userService, PollService pollService, StockService stockService) {
         this.i18n = i18n;
         this.userService = userService;
@@ -82,7 +80,7 @@ public class BookView extends CustomComponent implements View {
             return;
         }
         // Lookup
-        Stock stock = stockService.get(stockId);
+        var stock = stockService.get(stockId);
         if (stock == null) {
             ErrorView.show(String.format(i18n.get("book.error.missing"), stockId), null);
             return;
@@ -97,16 +95,16 @@ public class BookView extends CustomComponent implements View {
     }
 
     private Layout createBookLayout(Stock book) {
-        ContentPane contentPane = new ContentPane();
+        var contentPane = new ContentPane();
 
-        VisibilityIndicator visibility = new VisibilityIndicator(book.getVisibility(), i18n);
+        var visibility = new VisibilityIndicator(book.getVisibility(), i18n);
         contentPane.addComponent(visibility);
         contentPane.setComponentAlignment(visibility, Alignment.MIDDLE_RIGHT);
 
-        Image cover = new Image("", new ExternalResource(book.getThumbnailOrPlaceholder()));
+        var cover = new Image("", new ExternalResource(book.getThumbnailOrPlaceholder()));
         contentPane.addComponent(cover);
 
-        FormLayout form = new FormLayout();
+        var form = new FormLayout();
         form.setMargin(false);
         contentPane.addComponent(form);
 
@@ -116,76 +114,76 @@ public class BookView extends CustomComponent implements View {
             form.addComponent(owner);
         }
 
-        Label title = new Label(book.getTitle());
+        var title = new Label(book.getTitle());
         title.setCaption(i18n.get("book.column.title"));
         form.addComponent(title);
 
-        Label isbn = new Label(book.getIsbn());
+        var isbn = new Label(book.getIsbn());
         isbn.setCaption(i18n.get("book.column.isbn"));
         form.addComponent(isbn);
 
-        Label pageCount = new Label(String.valueOf(book.getPageCount()));
+        var pageCount = new Label(String.valueOf(book.getPageCount()));
         pageCount.setCaption(i18n.get("book.column.page.count"));
         form.addComponent(pageCount);
 
         if (!Strings.isNullOrEmpty(book.getPublishedDate())) {
-            Label publishedDate = new Label(book.getPublishedDate());
+            var publishedDate = new Label(book.getPublishedDate());
             publishedDate.setCaption(i18n.get("book.column.published.date"));
             form.addComponent(publishedDate);
         }
 
-        HorizontalLayout progress = new HorizontalLayout(new ProgressLabel(book, i18n));
+        var progress = new HorizontalLayout(new ProgressLabel(book, i18n));
         progress.setCaption(i18n.get("book.column.progress"));
         form.addComponent(progress);
 
         if (book.getProgress() == Stock.Progress.COMPLETED && book.getCompletedDate() != null) {
-            Label completedDate = new Label(DATE_FORMAT.format(book.getCompletedDate()));
+            var completedDate = new Label(DATE_FORMAT.format(book.getCompletedDate()));
             completedDate.setCaption(i18n.get("book.column.progress.date.completed"));
             form.addComponent(completedDate);
         }
 
-        Label obtainType = new Label(book.getObtainType().name());
+        var obtainType = new Label(book.getObtainType().name());
         obtainType.setCaption(i18n.get("book.column.obtain.type"));
         form.addComponent(obtainType);
 
         if (book.getObtainDate() != null) {
-            Label obtainDate = new Label(DATE_FORMAT.format(book.getObtainDate()));
+            var obtainDate = new Label(DATE_FORMAT.format(book.getObtainDate()));
             obtainDate.setCaption(i18n.get("book.column.obtain.date"));
             form.addComponent(obtainDate);
         }
 
         if (!Strings.isNullOrEmpty(book.getBoughtPlace())) {
-            Label boughtPlace = new Label(book.getBoughtPlace());
+            var boughtPlace = new Label(book.getBoughtPlace());
             boughtPlace.setCaption(i18n.get("book.column.bought.place"));
             form.addComponent(boughtPlace);
         }
 
-        String commentContent = book.getShortText();
-        Label comment = new Label();
+        var commentContent = book.getShortText();
+        var comment = new Label();
         comment.setCaption(i18n.get("book.column.text.short"));
         comment.setValue(Strings.isNullOrEmpty(commentContent) ?
                 i18n.get("book.label.text.short.empty") : commentContent);
         form.addComponent(comment);
 
-        String reviewContent = book.getLongText();
-        Label review = new Label("", ContentMode.HTML);
+        var reviewContent = book.getLongText();
+        var review = new Label("", ContentMode.HTML);
         review.setCaption(i18n.get("book.column.text.long"));
         review.setValue(Strings.isNullOrEmpty(reviewContent) ?
                 i18n.get("book.label.text.long.empty") : new PegDownProcessor().markdownToHtml(reviewContent));
         form.addComponent(review);
 
         if (userService.isLoggedIn()) {
-            Button backButton = new Button(i18n.get("books.label.title"),
+            var backButton = new Button(i18n.get("books.label.title"),
                     e -> getUI().getNavigator().navigateTo(BooksView.VIEW_NAME));
 
-            Button editButton = new Button(i18n.get("book.button.edit"), e -> {
+            var editButton = new Button(i18n.get("book.button.edit"), e -> {
                 VaadinSession.getCurrent().setAttribute(EditBookView.ATTR_BOOK_EDIT, book);
                 getUI().getNavigator().navigateTo(EditBookView.VIEW_NAME);
             });
             editButton.setIcon(VaadinIcons.EDIT);
             editButton.setStyleName(ValoTheme.BUTTON_PRIMARY);
 
-            HorizontalLayout buttonArea = new HorizontalLayout(backButton, editButton);
+            var buttonArea = new HorizontalLayout(backButton, editButton);
             buttonArea.setSpacing(true);
             contentPane.addComponent(buttonArea);
         }
