@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2017 Alice on Sunday Nights Workshop Participants. All rights reserved.
+ * Copyright (C) 2017-2018 Alice on Sunday Nights Workshop Participants. All rights reserved.
  */
 package io.github.aosn.mosaic.ui.view;
 
@@ -11,7 +11,6 @@ import com.vaadin.spring.annotation.SpringView;
 import com.vaadin.ui.*;
 import com.vaadin.ui.themes.ValoTheme;
 import io.github.aosn.mosaic.MosaicApplication;
-import io.github.aosn.mosaic.domain.model.stock.Stock;
 import io.github.aosn.mosaic.domain.service.auth.UserService;
 import io.github.aosn.mosaic.domain.service.poll.PollService;
 import io.github.aosn.mosaic.domain.service.stock.StockService;
@@ -20,10 +19,8 @@ import io.github.aosn.mosaic.ui.view.component.StockedBookTable;
 import io.github.aosn.mosaic.ui.view.layout.ContentPane;
 import io.github.aosn.mosaic.ui.view.layout.ViewRoot;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.vaadin.spring.i18n.I18N;
 
-import java.util.List;
 import java.util.stream.Collectors;
 
 /**
@@ -43,7 +40,6 @@ public class BooksView extends CustomComponent implements View {
     private transient final PollService pollService;
     private transient final StockService stockService;
 
-    @Autowired
     public BooksView(I18N i18n, UserService userService, PollService pollService, StockService stockService) {
         this.i18n = i18n;
         this.userService = userService;
@@ -58,14 +54,14 @@ public class BooksView extends CustomComponent implements View {
     }
 
     private Layout createMyStockLayout() {
-        ContentPane contentPane = new ContentPane();
+        var contentPane = new ContentPane();
 
-        List<Stock> stocks = stockService.getAll(userService.getUser());
+        var stocks = stockService.getAll(userService.getUser());
 
-        Label titleLabel = new Label(i18n.get("books.label.title"));
+        var titleLabel = new Label(i18n.get("books.label.title"));
         contentPane.addComponent(titleLabel);
 
-        Button addBookButton = new Button(i18n.get("books.button.add"),
+        var addBookButton = new Button(i18n.get("books.button.add"),
                 e -> getUI().getNavigator().navigateTo(FindBookView.VIEW_NAME));
         addBookButton.setIcon(VaadinIcons.PLUS);
         addBookButton.setStyleName(ValoTheme.BUTTON_FRIENDLY);
@@ -73,14 +69,14 @@ public class BooksView extends CustomComponent implements View {
         contentPane.setComponentAlignment(addBookButton, Alignment.MIDDLE_RIGHT);
 
         if (stocks.isEmpty()) {
-            Label label = new Label(VaadinIcons.INFO_CIRCLE.getHtml() + " " +
+            var label = new Label(VaadinIcons.INFO_CIRCLE.getHtml() + " " +
                     i18n.get("books.label.empty"), ContentMode.HTML);
             contentPane.addComponent(label);
             contentPane.setComponentAlignment(label, Alignment.MIDDLE_CENTER);
         } else {
-            List<StockedBookTable.Row> rows = stocks.stream()
+            var rows = stocks.stream()
                     .map(s -> StockedBookTable.Row.from(s, i18n)).collect(Collectors.toList());
-            StockedBookTable table = new StockedBookTable(rows, i18n);
+            var table = new StockedBookTable(rows, i18n);
             contentPane.addComponent(table);
         }
 
