@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2016-2017 Alice on Sunday Nights Workshop Participants. All rights reserved.
+ * Copyright (C) 2016-2018 Alice on Sunday Nights Workshop Participants. All rights reserved.
  */
 package io.github.aosn.mosaic.ui.view;
 
@@ -10,8 +10,6 @@ import com.vaadin.spring.annotation.SpringView;
 import com.vaadin.ui.*;
 import com.vaadin.ui.themes.ValoTheme;
 import io.github.aosn.mosaic.MosaicApplication;
-import io.github.aosn.mosaic.domain.model.auth.User;
-import io.github.aosn.mosaic.domain.model.poll.Book;
 import io.github.aosn.mosaic.domain.model.poll.Poll;
 import io.github.aosn.mosaic.domain.service.auth.UserService;
 import io.github.aosn.mosaic.domain.service.issue.IssueService;
@@ -25,10 +23,8 @@ import io.github.aosn.mosaic.ui.view.layout.IconAndName;
 import io.github.aosn.mosaic.ui.view.layout.ViewRoot;
 import io.github.aosn.mosaic.ui.view.style.Notifications;
 import io.github.aosn.mosaic.ui.view.style.Style;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.vaadin.spring.i18n.I18N;
 
-import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
@@ -53,7 +49,6 @@ public class PollResultView extends CustomComponent implements View {
     private transient final IssueService issueService;
     private transient final NotificationService notificationService;
 
-    @Autowired
     public PollResultView(I18N i18n, UserService userService, PollService pollService, IssueService issueService,
                           NotificationService notificationService) {
         this.i18n = i18n;
@@ -81,7 +76,7 @@ public class PollResultView extends CustomComponent implements View {
         }
 
         // Search DB
-        Poll poll = ErrorView.showIfExceptionThrows(() -> issueService.resolveBooks(pollService.get(pollId)));
+        var poll = ErrorView.showIfExceptionThrows(() -> issueService.resolveBooks(pollService.get(pollId)));
         if (poll == null) {
             return;
         }
@@ -97,35 +92,35 @@ public class PollResultView extends CustomComponent implements View {
     }
 
     private Layout createResultLayout(Poll poll) {
-        ContentPane contentPane = new ContentPane();
+        var contentPane = new ContentPane();
 
         contentPane.addComponent(new HeadingLabel(i18n.get("result.label.subject.prefix") + " " +
                 poll.getSubject()));
 
-        FormLayout aboutForm = new FormLayout();
+        var aboutForm = new FormLayout();
         aboutForm.setMargin(false);
         contentPane.addComponent(aboutForm);
 
-        IconAndName organizationLabel = new IconAndName(poll.getGroup());
+        var organizationLabel = new IconAndName(poll.getGroup());
         organizationLabel.setCaption(i18n.get("result.caption.poll.organization"));
         aboutForm.addComponent(organizationLabel);
 
-        IconAndName ownerLabel = new IconAndName(poll.getOwner());
+        var ownerLabel = new IconAndName(poll.getOwner());
         ownerLabel.setCaption(i18n.get("result.caption.poll.owner"));
         aboutForm.addComponent(ownerLabel);
 
-        String begin = poll.getBegin() == null ? "?" : PollTable.Row.DATE_FORMAT.format(poll.getBegin());
-        String end = poll.getEnd() == null ? "?" : PollTable.Row.DATE_FORMAT.format(poll.getEnd());
-        Label termLabel = new Label(begin + " - " + end);
+        var begin = poll.getBegin() == null ? "?" : PollTable.Row.DATE_FORMAT.format(poll.getBegin());
+        var end = poll.getEnd() == null ? "?" : PollTable.Row.DATE_FORMAT.format(poll.getEnd());
+        var termLabel = new Label(begin + " - " + end);
         termLabel.setCaption(i18n.get("result.caption.poll.term"));
         aboutForm.addComponent(termLabel);
 
-        List<User> users = poll.getVoters();
-        Label votesPerUserLabel = new Label(String.valueOf(users.size()));
+        var users = poll.getVoters();
+        var votesPerUserLabel = new Label(String.valueOf(users.size()));
         votesPerUserLabel.setCaption(i18n.get("result.caption.poll.voters.n"));
         aboutForm.addComponent(votesPerUserLabel);
 
-        CssLayout voters = new CssLayout();
+        var voters = new CssLayout();
         voters.setStyleName(Style.USER_COLLECTION.className());
         voters.setCaption(i18n.get("result.caption.poll.voters.list"));
         users.forEach(u -> voters.addComponent(new IconAndName(u)));
@@ -133,21 +128,21 @@ public class PollResultView extends CustomComponent implements View {
 
         // Show winner when poll was closed
         if (poll.getState() == Poll.PollState.CLOSED) {
-            Book winner = poll.getWinBook();
-            Label winnerLabel = new Label(winner == null ? i18n.get("result.label.poll.winner.tie") :
+            var winner = poll.getWinBook();
+            var winnerLabel = new Label(winner == null ? i18n.get("result.label.poll.winner.tie") :
                     winner.getGitHubIssue().getTitle());
             winnerLabel.setCaption(i18n.get("result.caption.poll.winner"));
             aboutForm.addComponent(winnerLabel);
 
             // Calculate popularity rate
             if (winner != null) {
-                Label popularityRateLabel = new Label(poll.calcPopularityRate(winner).toString());
+                var popularityRateLabel = new Label(poll.calcPopularityRate(winner).toString());
                 popularityRateLabel.setCaption(i18n.get("result.caption.popularity"));
                 aboutForm.addComponent(popularityRateLabel);
             }
         }
 
-        List<IssueTable.Row> rows = poll.getBooks().stream()
+        var rows = poll.getBooks().stream()
                 .map(r -> IssueTable.Row.from(r,
                         l -> issueService.isIssueLabel(l, poll.getGroup()),
                         l -> issueService.trimPartLabel(l, poll.getGroup())))
@@ -155,7 +150,7 @@ public class PollResultView extends CustomComponent implements View {
         contentPane.addComponent(new IssueTable(i18n.get("result.caption.book.list"), ColumnGroup.CLOSED, rows,
                 i18n));
 
-        Button backButton = new Button(i18n.get("common.button.back"),
+        var backButton = new Button(i18n.get("common.button.back"),
                 e -> getUI().getPage().setLocation(MainUI.PATH));
         contentPane.addComponent(backButton);
         contentPane.setComponentAlignment(backButton, Alignment.MIDDLE_CENTER);
@@ -164,31 +159,31 @@ public class PollResultView extends CustomComponent implements View {
         if (poll.isOwner(userService.getUser())) {
             contentPane.addComponent(new HeadingLabel(i18n.get("result.label.owner.operation")));
 
-            FormLayout summaryForm = new FormLayout();
+            var summaryForm = new FormLayout();
             summaryForm.setMargin(false);
             contentPane.addComponent(summaryForm);
 
-            Book winner = poll.judgeWinner();
-            String winnerName = winner == null ? i18n.get("result.label.poll.winner.tie") :
+            var winner = poll.judgeWinner();
+            var winnerName = winner == null ? i18n.get("result.label.poll.winner.tie") :
                     winner.getGitHubIssue().getTitle();
-            Label winnerLabel = new Label(winnerName);
+            var winnerLabel = new Label(winnerName);
             winnerLabel.setCaption(i18n.get("result.label.poll.winner.current"));
             summaryForm.addComponent(winnerLabel);
 
             if (winner != null) {
-                Label popularityRateLabel = new Label(poll.calcPopularityRate(winner).toString());
+                var popularityRateLabel = new Label(poll.calcPopularityRate(winner).toString());
                 popularityRateLabel.setCaption(i18n.get("result.caption.popularity"));
                 summaryForm.addComponent(popularityRateLabel);
             }
 
-            CheckBox notifyCheck = new CheckBox(i18n.get("common.caption.notify.slack"));
+            var notifyCheck = new CheckBox(i18n.get("common.caption.notify.slack"));
             notifyCheck.setValue(poll.getGroup().isSlackEnabled());
             notifyCheck.setEnabled(poll.getGroup().isSlackEnabled());
             contentPane.addComponent(notifyCheck);
 
-            String confirmMessage = i18n.get("result.label.confirm.close") + "<br/>" +
+            var confirmMessage = i18n.get("result.label.confirm.close") + "<br/>" +
                     i18n.get("result.label.poll.winner.current") + ": " + winnerName;
-            Button closeButton = new Button(i18n.get("result.button.poll.close"),
+            var closeButton = new Button(i18n.get("result.button.poll.close"),
                     e -> UI.getCurrent().addWindow(new ConfirmWindow(confirmMessage, i18n, ok -> {
                         pollService.close(poll);
                         if (notifyCheck.getValue()) {
